requirements.txt
fastapi
jinja2
uvicorn[standard]
load_dotenv
arel